Code PlaceOrderServlet and OrderServices ( code placeOrder method in OrderServices)

+++ BookStoreWebsite/src/com/bookstore/controller/frontend/shoppingcart/ShoppingCart.java
@@ -38,15 +38,15 @@ public class ShoppingCart {
 		return total;
 	}
 
-	public double getTotalAmount() {
-		double total = 0.0f;
+	public float getTotalAmount() {
+		float total = 0.0f;
 
 		Iterator<Book> iterator = cart.keySet().iterator();
 
 		while (iterator.hasNext()) {
 			Book book = iterator.next();
 			Integer quantity = cart.get(book); // Returns the value to which the specified key is mapped
-			double subTotal = quantity * book.getPrice(); //
+			float subTotal = quantity * book.getPrice(); //
 			total += subTotal;
 		}
